+++ src/Lansweeper.Heijden.Dns/Enums/TransportType.cs
@@ -1,0 +1,13 @@
+/*
+ * http://www.iana.org/assignments/dns-parameters
+ *
+ *
+ *
+ */
+namespace Heijden.Dns.Enums;
+
+public enum TransportType
+{
+    Udp,
+    Tcp
+}
