+++ src/Lansweeper.Heijden.Dns/Records/Record.cs
@@ -7,6 +7,6 @@ namespace Heijden.DNS
 		/// <summary>
 		/// The Resource Record this RDATA record belongs to
 		/// </summary>
-		public RR RR;
+		public RR? RR { get; set; }
 	}
 }
